+++ Sailock/Views/GeneratorView.xaml.cs
@@ -18,5 +18,10 @@ namespace Sailock.Views
         {
 
         }
+
+        private void TextBox_TextChanged_1(object sender, TextChangedEventArgs e)
+        {
+
+        }
     }
 }